added file_edit to the new site system
